requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
flake8==5.0.4
flake8-docstrings==1.6.0
pyTelegramBotAPI==4.14.1
python-dotenv==0.20.0